refactor: 패키지명 reservation 으로 변경
