+++ src/main/java/roomescape/reservation/JdbcReservationRepository.java
@@ -1,4 +1,4 @@
-package roomescape.entity;
+package roomescape.reservation;
 
 import java.util.List;
 import org.springframework.jdbc.core.JdbcTemplate;
