+++ src/main/java/roomescape/reservation/ReservationRepository.java
@@ -1,4 +1,4 @@
-package roomescape.entity;
+package roomescape.reservation;
 
 import java.util.List;
