+++ src/test/java/roomescape/reservation/ReservationTest.java
@@ -1,4 +1,4 @@
-package roomescape.entity;
+package roomescape.reservation;
 
 import static org.assertj.core.api.Assertions.assertThat;
