+++ src/test/java/roomescape/reservation/ReservationsTest.java
@@ -1,4 +1,4 @@
-package roomescape.entity;
+package roomescape.reservation;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
